+++ InscryptionCommunityPatch/Card/Vanilla Ability Patches/Act2SubmergeFix.cs
@@ -1,17 +1,15 @@
 using DiskCardGame;
 using HarmonyLib;
-using InscryptionAPI.Card;
 using System.Collections;
 using UnityEngine;
 
 namespace InscryptionCommunityPatch.Card;
 
-[HarmonyPatch]
+[HarmonyPatch(typeof(Submerge), nameof(Submerge.OnUpkeep))]
 internal class Act2SubmergeFix
 {
-    [HarmonyPatch(typeof(Submerge), nameof(Submerge.OnUpkeep))]
     [HarmonyPostfix]
-    private static IEnumerator AlwaysResolveOnBoard(IEnumerator enumerator)
+    private static IEnumerator AddWaitInAct2(IEnumerator enumerator)
     {
         yield return enumerator;
         if (SaveManager.SaveFile.IsPart2)